web-server/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic-settings
pydantic[email]
alembic
SQLAlchemy
python-multipart
python-jose
passlib
openai
opencv-python-headless
google-cloud-storage